renames 'ark' to 'client'
