+++ .circleci/config.yml
@@ -16,7 +16,7 @@ jobs:
           name: run tests
           command: |
             . venv/bin/activate
-            pytest -v -s --cov=ark --cov-config=.coveragerc --cov-report xml
+            pytest -v -s --cov=client --cov-config=.coveragerc --cov-report xml
       - run:
           name: Codecov
           when: on_success
