+++ .coveragerc
@@ -1,6 +1,6 @@
 [run]
 branch = True
-source = ark
+source = client
 
 [report]
 exclude_lines =
