+++ client/api/two/peers.py
@@ -1,4 +1,4 @@
-from ark.resource import Resource
+from client.resource import Resource
 
 
 class Peers(Resource):
