+++ client/exceptions.py
