+++ setup.cfg
@@ -3,7 +3,7 @@ test=pytest
 
 [flake8]
 max-line-length = 100
-application-import-names = ark
+application-import-names = client
 
 [tool:pytest]
 addopts = --verbose -s
