+++ tests/api/two/test_peers.py
@@ -1,6 +1,6 @@
 import responses
 
-from ark import ArkClient
+from client import ArkClient
 
 
 def test_all_calls_correct_url_with_default_params():
